modified.

+++ mpi_assignment_2.py
@@ -3,7 +3,7 @@ from mpi4py import MPI
 comm = MPI.COMM_WORLD
 rank = comm.Get_rank()
 
-randNum = numpy.zeros(1,dtype = numpy.int)
+randNum = numpy.zeros(1)
 
         
 if rank == 0:
